// src/models/Person.d.ts
export interface InputPerson {
  name: number;
  r: number;
  f: number;
  of: number;
  or: number;
  c: number;
  steam: string;
  xbox: string;
  psn: string;
}

export interface Person {
  name: number;
  rank: number;
  fame: number;
  original_fame: number;
  original_rank: number;
  cashout: number;
  steam: string;
  xbox: string;
  psn: string;
}